Refactor to separate classes in their specific files an tests improuvements for styles
