+++ src/file/numbering/numbering.spec.ts
@@ -25,23 +25,21 @@ describe("Numbering", () => {
                 { "w:multiLevelType": [{ _attr: { "w:val": "hybridMultilevel" } }] },
             ]);
 
-            abstractNums
-                .filter((el) => el["w:lvl"])
-                .forEach((el, ix) => {
-                    expect(Object.keys(el)).to.have.lengthOf(1);
-                    expect(Object.keys(el["w:lvl"]).sort()).to.deep.equal(["_attr", "w:start", "w:lvlJc", "w:numFmt", "w:pPr", "w:rPr"]);
-                    expect(el["w:lvl"]).to.have.deep.members([
-                        { _attr: { "w:ilvl": ix, "w15:tentative": 1 } },
-                        { "w:start": [{ _attr: { "w:val": 1 } }] },
-                        { "w:lvlJc": [{ _attr: { "w:val": "left" } }] },
-                        { "w:numFmt": [{ _attr: { "w:val": "bullet" } }] },
-                    ]);
-                    // Once chai 4.0.0 lands and #644 is resolved, we can add the following to the test:
-                    // {"w:lvlText": [{"_attr": {"w:val": "•"}}]},
-                    // {"w:rPr": [{"w:rFonts": [{"_attr": {"w:ascii": "Symbol", "w:cs": "Symbol", "w:eastAsia": "Symbol", "w:hAnsi": "Symbol", "w:hint": "default"}}]}]},
-                    // {"w:pPr": [{"_attr": {}},
-                    //            {"w:ind": [{"_attr": {"w:left": 720, "w:hanging": 360}}]}]},
-                });
+            abstractNums.filter((el) => el["w:lvl"]).forEach((el, ix) => {
+                expect(Object.keys(el)).to.have.lengthOf(1);
+                expect(Object.keys(el["w:lvl"]).sort()).to.deep.equal(["_attr", "w:start", "w:lvlJc", "w:numFmt", "w:pPr", "w:rPr"]);
+                expect(el["w:lvl"]).to.have.deep.members([
+                    { _attr: { "w:ilvl": ix, "w15:tentative": 1 } },
+                    { "w:start": [{ _attr: { "w:val": 1 } }] },
+                    { "w:lvlJc": [{ _attr: { "w:val": "left" } }] },
+                    { "w:numFmt": [{ _attr: { "w:val": "bullet" } }] },
+                ]);
+                // Once chai 4.0.0 lands and #644 is resolved, we can add the following to the test:
+                // {"w:lvlText": [{"_attr": {"w:val": "•"}}]},
+                // {"w:rPr": [{"w:rFonts": [{"_attr": {"w:ascii": "Symbol", "w:cs": "Symbol", "w:eastAsia": "Symbol", "w:hAnsi": "Symbol", "w:hint": "default"}}]}]},
+                // {"w:pPr": [{"_attr": {}},
+                //            {"w:ind": [{"_attr": {"w:left": 720, "w:hanging": 360}}]}]},
+            });
         });
     });
